# openapi.yml
openapi: 3.0.4
info:
  title: User Service API
  version: 2.0.0
  description: >
    Сервис управления пользователями и аутентификацией. Обрабатывает регистрацию,
    вход, обновление токенов, выход, настройки аккаунта и заявки на роль преподавателя.
servers:
  - url: http://localhost:8081
    description: Local Dev Server
...
paths:
...
  /api/v1/users/me:
    get:
      summary: Получить профиль текущего пользователя
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Access токен отсутствует, невалидный или истек
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Full authentication is required"
                code: 401
                timestamp: "2026-02-24T12:00:00Z"
        '500':
          description: Ошибка сервиса
  /api/v1/users/me/settings:
    get:
      summary: Получить настройки текущего пользователя
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Настройки аккаунта пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettingsResponse'
        '401':
          description: Access токен отсутствует, невалидный или истек
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/me/password:
    put:
      summary: Сменить пароль текущего пользователя
      tags: [Users]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Пароль успешно обновлен
        '400':
          description: Ошибка валидации запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Access токен отсутствует, невалидный или истек
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Обновить настройки текущего пользователя
      tags: [Users]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSettingsRequest'
      responses:
        '200':
          description: Настройки аккаунта обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettingsResponse'
        '400':
          description: Ошибка валидации запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Access токен отсутствует, невалидный или истек
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/users/{id}:
    get:
      summary: Получить пользователя по ID (только для администраторов)
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID пользователя
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Access токен отсутствует, невалидный или истек
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Full authentication is required"
                code: 401
                timestamp: "2026-02-24T12:00:00Z"
        '403':
          description: Недостаточно прав (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Access Denied"
                code: 403
                timestamp: "2026-02-24T12:00:00Z"
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "User not found"
                code: 404
                timestamp: "2026-02-24T12:00:00Z"
        '500':
          description: Ошибка сервиса
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT токен, полученный при аутентификации через /auth/login.
        Токен должен быть передан в заголовке Authorization: Bearer <token>
  schemas:
...
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          example: "securePassword123"
        newPassword:
          type: string
          minLength: 8
          example: "newSecurePassword123"
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID пользователя
          example: 2
        email:
          type: string
          description: Email пользователя
          example: "test2@mail.com"
        fullName:
          type: string
          description: Полное имя пользователя
          example: "Тест Тестов"
        role:
          type: string
          enum: [STUDENT, TEACHER, ADMIN]
          description: Роль пользователя в системе
          example: "STUDENT"
        status:
          type: string
          enum: [ACTIVE, BLOCKED, DELETED]
          description: Статус пользователя
          example: "ACTIVE"
        avatarUrl:
          type: string
          nullable: true
          maxLength: 1000
          example: "https://example.com/avatar.png"
        bio:
          type: string
          nullable: true
          maxLength: 2000
          example: "Java learner"
        preferredLocale:
          type: string
          enum: [ru, en]
          example: "ru"
    UserSettingsRequest:
      type: object
      required: [fullName, preferredLocale]
      properties:
        fullName:
          type: string
          maxLength: 255
          example: "Roman Aksenov"
        avatarUrl:
          type: string
          nullable: true
          maxLength: 1000
          example: "https://example.com/avatar.png"
        bio:
          type: string
          nullable: true
          maxLength: 2000
          example: "Java learner"
        preferredLocale:
          type: string
          enum: [ru, en]
          example: "ru"
    UserSettingsResponse:
      allOf:
        - $ref: '#/components/schemas/UserDto'
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Описание ошибки
        code:
          type: integer
          description: HTTP код ошибки
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки